menu order, def position in common.cpp

+++ car/src/common.hpp
@@ -11,4 +11,4 @@ const std::string default_profile = "/etc/air/profile";
 // NOLINTNEXTLINE(cppcoreguidelines-avoid-non-const-global-variables)
 extern profile car_profile;
 
-extern int position;
+extern uint8_t position;
